+++ core-ng/src/main/java/core/framework/internal/stat/CPUStat.java
@@ -9,15 +9,12 @@ class CPUStat {
     private final OperatingSystemMXBean os;
     private final int availableProcessors;
 
-    private long previousCPUTime;
-    private long previousTime;
+    private long previousCPUTime = -1;
+    private long previousTime = -1;
 
     CPUStat(OperatingSystemMXBean os) {
         this.os = os;
         this.availableProcessors = os.getAvailableProcessors();
-
-        previousTime = System.nanoTime();
-        previousCPUTime = os.getProcessCpuTime();
     }
 
     // due to os.getProcessCpuLoad() returns host level info, not for container, so here uses process time to calculate container level java cpu usage
@@ -38,6 +35,7 @@ class CPUStat {
     }
 
     double usage(long currentTime, long previousTime, long currentCPUTime, long previousCPUTime, int availableProcessors) {
+        if (previousCPUTime == -1) return 0;    // ignore first data point, to collect cpu usage after startup
         return (double) (currentCPUTime - previousCPUTime) / (currentTime - previousTime) / availableProcessors;
     }
 }